refactor: seaprate env check logic to utils from config
