+++ apps/web/src/shared/utils/index.ts
@@ -1,2 +1,3 @@
 ﻿export * from './storage';
 export * from './api';
+export * from './env';
